+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## 0.6.2 (2019-10-14)
 - added `demo.ipynb`
 - self-contained `README.md`
+- Fixed piling up Kibana objects by introducing `ifexists='reuse_existing'`
 
 ## 0.6.1 (2019-10-11)
 - Package infrastructure: pypi, travis-CI, pytest